Correctif sur des initialisations de variables.

+++ mod_engagements/saisie_engagements.php
@@ -290,6 +290,8 @@ if((isset($id_classe))&&(isset($_POST['is_posted']))&&($engagement_statut=='elev
 			if($acces_saisie=="y") {
 
 				for($loop2=0;$loop2<count($current_login);$loop2++) {
+					// On parcourt les élèves et les parents, mais avec get_info_eleve(), on ne récupère ensuite que les élèves.
+
 					$chaine=$id_classe[$loop]."|".$current_login[$loop2]."|".$current_id_engagement;
 					//echo "$chaine<br />";
 					//if(!in_array($chaine, $engagement)) {
@@ -300,7 +302,7 @@ if((isset($id_classe))&&(isset($_POST['is_posted']))&&($engagement_statut=='elev
 					print_r($tmp_info_user);
 					echo "</pre>";
 					*/
-					if((!in_array($chaine, $engagement))&&($tmp_info_user['statut']=='eleve')) {
+					if((!in_array($chaine, $engagement))&&(isset($tmp_info_user['statut']))&&($tmp_info_user['statut']=='eleve')) {
 						$sql="DELETE FROM engagements_user WHERE login='".$current_login[$loop2]."' AND id_type='id_classe' AND valeur='".$id_classe[$loop]."' AND id_engagement='$current_id_engagement';";
 						//echo "$sql<br />";
 						$del=mysqli_query($GLOBALS["mysqli"], $sql);
@@ -452,13 +454,20 @@ if((isset($id_classe))&&(isset($_POST['is_posted']))&&($engagement_statut=='resp
 
 			if($acces_saisie=="y") {
 				for($loop2=0;$loop2<count($current_login);$loop2++) {
+					// On parcourt les élèves et les parents, mais avec get_info_responsable(), on ne récupère ensuite que les parents.
+
 					$chaine=$id_classe[$loop]."|".$current_login[$loop2]."|".$current_id_engagement;
 
 					//echo "$chaine<br />";
 					// Il ne faut pas désinscrire les élèves ici
 					//$tmp_info_user=get_info_user($current_login[$loop2]);
 					$tmp_info_user=get_info_responsable($current_login[$loop2]);
-					if((!in_array($chaine, $engagement))&&($tmp_info_user['statut']=='responsable')) {
+					/*
+					echo "<pre>";
+					print_r($tmp_info_user);
+					echo "</pre>";
+					*/
+					if((!in_array($chaine, $engagement))&&(isset($tmp_info_user['statut']))&&($tmp_info_user['statut']=='responsable')) {
 						$sql="DELETE FROM engagements_user WHERE login='".$current_login[$loop2]."' AND id_type='id_classe' AND valeur='".$id_classe[$loop]."' AND id_engagement='$current_id_engagement';";
 						//echo "$sql<br />";
 						$del=mysqli_query($GLOBALS["mysqli"], $sql);
